Expose PGPSecretKeyRing constructor
The PGPSecretKeyRing is in its essence a simple enough wrapper around a
list of PGPSecretKeys. It can however only be instantiated from a binary
representation, a PGPKeyRingGenerator or bootstrapped by modifying an
existing SecretKeyRing. This is comonly fine ensure constraints from the
constructor are preserved -- but we need the access in OpenKeychain
because we just work on that low level.

Hrum.